src/requirements.txt
rcon
twitch-python